bootloader: kernel download from python script and working
Known issues are related to the compiler optimization level. When
switching to kernel image on -O0 the PC does not seem to get updated. I
have checked the memory and everything looks fine.

The second issue is related to serial printing after reboot. A wird `y`
character gets printed instead.